pyproject.toml
[project]
name = "sofia"
version = "0.1.0"
description = "Sofia — AI Coaching Assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "PyJWT>=2.7.0",
    "bcrypt>=4.0.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "chromadb>=0.5.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=4.0.0",
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
